Module cold_iron::detect

Library for building Thaumatic Intrusion Detection Systems

One of the very first uses of warding was intrusion detection. The witches of old would draw glyphs and spell circles on their doors and windows, alerting them of any intruders of malevolent intent.

It is a relatively recent discovery that this protection and disclosure is not limited merely to the physical realm. From a thaumaturgic perspective, there is little difference between a cottage and a core network switch. And just as a spell circle won’t exclude a close friend or beloved pet, a well-configured thaumatic firewall will allow approved users and machines unimpeded access.

The field of Thaumatic Intrusion Detection Systems (TIDS) is burgeoning, and certainly out of the scope of a single library. However, the primitives described here are sufficiently expressive to protect a personal computer or home server.

The most difficult part of configuring a TIDS is defining the boundary, and clearly picturing what should and shouldn’t be allowed. Especially when starting out, it’s a good idea to rigorously test any newly created TIDS to ensure no loopholes exist. And remember that no firewall is fool proof: a counterthaumatist of sufficient skill will be able to step inside all but the most thoroughly secured abode.

Structs

Ward
Ward represents a watcher for some entity. It must be assigned to a network address or location

Enums

MalevolentEntity