Module cold_iron::net

Libraries for long-distance thaumatic communication

Warning! You need a license to set up a long-distance Thaum station! Please enable the feature i_promise_i_am_licensed to confirm you’re registered with your local thaumatic authority! Failure to do so could result in civil penalties!

Whether spell circles can actually be used to commune with the dead or with demons is still hotly debated. But we know for certain that they can be used to communicate with each other. One of the first applications of nanothaumaturgy was long-distance communication, without the need for wires or radio waves.

While the distance of thaumatic communication seems to be nearly unlimited, the bandwidth is unfortunately quite small. A modern SITA under ideal conditions would expect around 1 megabit per second, full simplex. So while nanothaumaturgy has begun to supplant Iridium and other satellite phone providers, it is unlikely to replace Ethernet or Wifi anytime soon.

Unlike the complicated hierarchies of TCP/IP, thaumatic communications have a completely flat topology. Any spell circle can reach any other circle, provided both are listening in the right places. To prevent collisions and confusion, addresses may be requested from the International Runic Address Authority (IRAA), sister organisation to the IANA.

To distinguish runic addresses from IP addresses, runes are used instead of numbers. These are not merely cosmetic or Direction: The simple geometric primitives are embedded directly into the Shape of a thaumaturgic incantation. The incantation also specifies runic addresses to allow connections from. It is possible, though discouraged, to create an “open circle” that allows connections from anywhere.

Structs

RunicAddress

ThaumListener

ThaumSocket

ThaumStream